+++ app/database/migrations/2014_07_31_081207_create_production_accessories.php
@@ -12,7 +12,7 @@ class CreateProductionAccessories extends Migration {
             Schema::create($this->table, function(Blueprint $table) {
                 $table->increments('id');
                 $table->string('title',128)->nullable();
-                $table->float('price')->default(0)->nullable();
+                $table->double('price')->default(0)->nullable();
                 $table->text('description')->nullable();
                 $table->integer('category_id')->default(0)->unsigned()->nullable();
                 $table->integer('accessibility_id')->default(0)->unsigned()->nullable();
